+++ changelog.md
@@ -7,6 +7,7 @@
 - Fix crusher not accepting multiple stacks of the same ore (BluSunrize)
 - Fix stone multiblocks being washed away by water (BluSunrize)
 - Fix potential race condition with wires damaging entities (BluSunrize)
+- Fix shader banners and potion fluids using the wrong colors (BluSunrize)
 
 ##### Version 1.20.4-11.3.0-174
 - Re-add CC-Tweaked integration (Malte)